% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lavaan.R
\name{fit_lavaan}
\alias{fit_lavaan}
\title{Get the lavaan fit indexes in tidy format}
\usage{
fit_lavaan(x)
}
\arguments{
\item{x}{A lavaan object that is estimated by \code{\link[lavaan:cfa]{lavaan::cfa()}} or
\code{\link[lavaan:sem]{lavaan::sem()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with a row and each estimated parameter in a
column

\item{nobs}{Number of analysed observations}
\item{estimator}{Estimator used}
\item{ngroups}{Number of groups in model}
\item{converged}{Logical indicator about the convergence of the model}
\item{chisq}{Model chi squared}
\item{df}{degrees of freedom}
\item{pvalue}{P-value associated with the model}
\item{npar}{Number of parameters in the model}
\item{CFI}{Comparative Fit Index}
\item{TLI}{Tucker-Lewis Index also known as NNFI}
\item{RMSEA}{Root Mean Square error of Approximation}
\item{RMSEA.CI.LOWER}{95 percent lower bound on RMSEA}
\item{RMSEA.CI.UPPER}{95 percent upper bound on RMSEA}
\item{SRMR}{Standardized Root Mean squared Residual}
\item{WRMR}{Weighted Root Mean Square Residual}
\item{AIC}{Akaike information criterion}
\item{BIC}{Bayesian information criterion}
\item{missing_method}{Method for eliminating missing data}
}
\description{
\code{fit_lavaan} allows to extract the fit indexes of a factorial
model estimated with lavaan. The main feature of this function is that, if
the estimator is robust, the fit indexes obtained will be the robust ones.
}
\examples{

library(lavaan)

HS.model <- " visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 "

fit <- cfa(
  model = HS.model,
  data = HolzingerSwineford1939
)

fit_lavaan(fit)
}
\seealso{
\code{\link[lavaan:cfa]{lavaan::cfa()}}, \code{\link[lavaan:sem]{lavaan::sem()}}, \code{\link[lavaan:fitMeasures]{lavaan::fitmeasures()}}
}
