% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_invariance.R
\name{identify_items_partial}
\alias{identify_items_partial}
\title{Identify items for partial invariance}
\usage{
identify_items_partial(fit, cutoff.value = 0.1, detailed = FALSE)
}
\arguments{
\item{fit}{A lavaan object that is estimated by \code{\link[lavaan:cfa]{lavaan::cfa()}},
\code{\link[lavaan:sem]{lavaan::sem()}} or \code{\link[semTools:measEq.syntax]{semTools::measEq.syntax()}}.}

\item{cutoff.value}{By default, only p-values below 0.10 are retained.
If you want to obtain all of them in their entirety, indicate the value
as \code{NULL}.}

\item{detailed}{Defaults to FALSE, and will display only the essential
information requested. If you want the complete columns obtained with
\code{\link[lavaan:lavParTable]{lavaan::parTable()}}, set to TRUE.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} to identify items and specifications that,
when indicated, could solve non-invariance problems.
}
\description{
This function uses 2 lavaan functions to generate
a table (in tibble format) to identify which items and
specifications among the groups could be generating
problematic situations.
}
\examples{

library(semTools)
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit_scalar <- measEq.syntax(configural.model = HS.model,
                            data = lavaan::HolzingerSwineford1939,
                            estimator = "MLR",
                            parameterization = "theta",
                            ID.fac = "std.lv",
                            ID.cat = "Wu.Estabrook.2016",
                            group = "sex",
                            group.equal = c("loadings", "intercepts"),
                            return.fit = TRUE)

identify_items_partial(fit_scalar)

}
\seealso{
\code{\link[lavaan:lavParTable]{lavaan::parTable()}}, \code{\link[lavaan:lavTestScore]{lavaan::lavTestScore()}}
}
